@@ -18,7 +18,7 @@ import { MenubarModule } from 'primeng/menubar';
 import { InputTextModule } from 'primeng/inputtext';
 import { HeaderComponent } from './components/header/header.component';
 import { UploaderComponent } from './components/uploader/uploader.component';
-import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { BoundComponent } from './form/bound/bound.component';
 import { BoundColorsComponent } from './form/bound-colors/bound-colors.component';
 import { TabViewModule } from 'primeng/tabview';
@@ -104,133 +104,126 @@ import { ShopcartPrintItemComponent } from './components/shopcart-print-item/sho
 import { PrintTypeCoverComponent } from './form/print-type-cover/print-type-cover.component';
 import { PrintGuideComponent } from './components/print-guide/print-guide.component';
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    PaperSizeComponent,
-    PaperTypeComponent,
-    PrintTypeComponent,
-    PrintFormComponent,
-    PagesPerSideComponent,
-    OrientationComponent,
-    FinishTypeComponent,
-    BoundComponent,
-    HeaderComponent,
-    UploaderComponent,
-    BoundTypeComponent,
-    FilterPipe,
-    BoundColorsComponent,
-    ColorOptionComponent,
-    QuantityCopiesComponent,
-    AdditionalCommentComponent,
-    ConfirmBarComponent,
-    ShopcartWrapperComponent,
-    PrintComponent,
-    PaymentComponent,
-    ShopcartComponent,
-    TermsComponent,
-    SuccessComponent,
-    ErrorComponent,
-    LoadingComponent,
-    LoginComponent,
-    ShippingComponent,
-    BillingComponent,
-    ProfileComponent,
-    AdminIndexComponent,
-    BannerComponent,
-    ShopIndexComponent,
-    ProductComponent,
-    AdminShopComponent,
-    AdminBannerComponent,
-    PillsComponent,
-    FooterComponent,
-    CarouselComponent,
-    SignUpComponent,
-    ResetPasswordComponent,
-    ForgotPasswordComponent,
-    ShopProductComponent,
-    UserOrdersComponent,
-    UserInformationComponent,
-    OrderDetailsComponent,
-    PrintTypeCoverComponent,
-    UserPasswordComponent,
-    BoundPagesComponent,
-    OrderProcessingComponent,
-    PaymentsEmptyCartWarningComponent,
-    MenuSidebarComponent,
-    MobileMenuComponent,
-    ViewTarjetasVisitaComponent,
-    ViewFlyersComponent,
-    ViewCarpetasComponent,
-    ViewRevistasComponent,
-    ViewRollupsComponent,
-    ViewCartelesComponent,
-    ViewDipticosComponent,
-    ViewTripticosComponent,
-    InputCustomSizeComponent,
-    SelectButtonComponent,
-    ShopcartPrintItemComponent,
-    PrintGuideComponent,
-  ],
-  imports: [
-    BrowserModule,
-    DropdownModule,
-    FormsModule,
-    ReactiveFormsModule,
-    TabMenuModule,
-    AppRoutingModule,
-    FormsModule,
-    TableModule,
-    MenuModule,
-    DividerModule,
-    OverlayPanelModule,
-    DataViewModule,
-    InputSwitchModule,
-    SelectButtonModule,
-    FileUploadModule,
-    RadioButtonModule,
-    MenubarModule,
-    HttpClientModule,
-    InputTextModule,
-    InputNumberModule,
-    GalleriaModule,
-    TabViewModule,
-    CardModule,
-    ConfirmDialogModule,
-    CheckboxModule,
-    InputTextareaModule,
-    BadgeModule,
-    ToastModule,
-    BrowserAnimationsModule,
-    SidebarModule,
-    StoreModule.forRoot({
-      customer: customerReducer,
-      loading: loadingReducer,
-      coupons: couponsReducer,
-    }),
-    EffectsModule.forRoot([CustomerEffects]),
-  ],
-  providers: [
-    MessageService,
-    ConfirmationService,
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: AuthInterceptor,
-      multi: true,
-    },
-
-    {
-      provide: HTTP_INTERCEPTORS,
-      useClass: HttpErrorInterceptor,
-      multi: true,
-    },
-    {
-      provide: APP_INITIALIZER,
-      useFactory: (authService: AuthService) => initializeApp(authService),
-      multi: true,
-      deps: [AuthService],
-    },
-  ],
-  bootstrap: [AppComponent],
-})
+@NgModule({ declarations: [
+        AppComponent,
+        PaperSizeComponent,
+        PaperTypeComponent,
+        PrintTypeComponent,
+        PrintFormComponent,
+        PagesPerSideComponent,
+        OrientationComponent,
+        FinishTypeComponent,
+        BoundComponent,
+        HeaderComponent,
+        UploaderComponent,
+        BoundTypeComponent,
+        FilterPipe,
+        BoundColorsComponent,
+        ColorOptionComponent,
+        QuantityCopiesComponent,
+        AdditionalCommentComponent,
+        ConfirmBarComponent,
+        ShopcartWrapperComponent,
+        PrintComponent,
+        PaymentComponent,
+        ShopcartComponent,
+        TermsComponent,
+        SuccessComponent,
+        ErrorComponent,
+        LoadingComponent,
+        LoginComponent,
+        ShippingComponent,
+        BillingComponent,
+        ProfileComponent,
+        AdminIndexComponent,
+        BannerComponent,
+        ShopIndexComponent,
+        ProductComponent,
+        AdminShopComponent,
+        AdminBannerComponent,
+        PillsComponent,
+        FooterComponent,
+        CarouselComponent,
+        SignUpComponent,
+        ResetPasswordComponent,
+        ForgotPasswordComponent,
+        ShopProductComponent,
+        UserOrdersComponent,
+        UserInformationComponent,
+        OrderDetailsComponent,
+        PrintTypeCoverComponent,
+        UserPasswordComponent,
+        BoundPagesComponent,
+        OrderProcessingComponent,
+        PaymentsEmptyCartWarningComponent,
+        MenuSidebarComponent,
+        MobileMenuComponent,
+        ViewTarjetasVisitaComponent,
+        ViewFlyersComponent,
+        ViewCarpetasComponent,
+        ViewRevistasComponent,
+        ViewRollupsComponent,
+        ViewCartelesComponent,
+        ViewDipticosComponent,
+        ViewTripticosComponent,
+        InputCustomSizeComponent,
+        SelectButtonComponent,
+        ShopcartPrintItemComponent,
+        PrintGuideComponent,
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        DropdownModule,
+        FormsModule,
+        ReactiveFormsModule,
+        TabMenuModule,
+        AppRoutingModule,
+        FormsModule,
+        TableModule,
+        MenuModule,
+        DividerModule,
+        OverlayPanelModule,
+        DataViewModule,
+        InputSwitchModule,
+        SelectButtonModule,
+        FileUploadModule,
+        RadioButtonModule,
+        MenubarModule,
+        InputTextModule,
+        InputNumberModule,
+        GalleriaModule,
+        TabViewModule,
+        CardModule,
+        ConfirmDialogModule,
+        CheckboxModule,
+        InputTextareaModule,
+        BadgeModule,
+        ToastModule,
+        BrowserAnimationsModule,
+        SidebarModule,
+        StoreModule.forRoot({
+            customer: customerReducer,
+            loading: loadingReducer,
+            coupons: couponsReducer,
+        }),
+        EffectsModule.forRoot([CustomerEffects])], providers: [
+        MessageService,
+        ConfirmationService,
+        {
+            provide: HTTP_INTERCEPTORS,
+            useClass: AuthInterceptor,
+            multi: true,
+        },
+        {
+            provide: HTTP_INTERCEPTORS,
+            useClass: HttpErrorInterceptor,
+            multi: true,
+        },
+        {
+            provide: APP_INITIALIZER,
+            useFactory: (authService: AuthService) => initializeApp(authService),
+            multi: true,
+            deps: [AuthService],
+        },
+        provideHttpClient(withInterceptorsFromDi()),
+    ] })
 export class AppModule {}
